Make `PrivateHeaders` importable
Summary:
## This stack

For DevInfra AI Native Week, I wanted to try and modernize the `idb` codebase and convert it to Swift. This stack is that project. It does not touch anything in the original `idb` codebase or `idb-cli` and is not used in production anywhere. Because the changes are so sweeping, I plan to aggressively test the finished product to validate the refactoring is a success.

## This diff

Fix the class-dumped frameworks so they are importable into Swift code.

The class-dumped headers are missing module maps and contain all kinds of weirdness (like their own definitions for Foundation structs). These issues need to be fixed for them to import correctly into Swift code.

## Session

DEV138537427

```
claude --resume 008dadcb-b612-45d9-89a4-b08eb895cd37
```

Reviewed By: bryanyantx

Differential Revision: D98734691

fbshipit-source-id: 41af0415b0bbceea6ba4dd162e190fe10f05da7f

+++ PrivateHeaders/CoreSimulator/DefaultDisplayDescriptorState.h
@@ -12,7 +12,7 @@
 @interface DefaultDisplayDescriptorState : NSObject <SimDisplayDescriptorState>
 {
   int _powerState;
-  int _displayClass;
+  unsigned short _displayClass;
   unsigned int _defaultWidthForDisplay;
   unsigned int _defaultHeightForDisplay;
 }
@@ -20,7 +20,7 @@
 + (id)defaultDisplayDescriptorStateWithPowerState:(int)arg1 displayClass:(int)arg2 width:(unsigned int)arg3 height:(unsigned int)arg4;
 @property (nonatomic, assign) unsigned int defaultHeightForDisplay;
 @property (nonatomic, assign) unsigned int defaultWidthForDisplay;
-@property (nonatomic, assign) int displayClass;
+@property (nonatomic, assign) unsigned short displayClass;
 @property (nonatomic, assign) int powerState;
 - (id)xpcObject;
